requirements.txt
amqp==2.4.2
billiard==3.5.0.5
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.9.2
djoser==1.5.0
kombu==4.4.0
pendulum==2.0.4
psycopg2==2.7.7
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
pytzdata==2018.9
six==1.12.0
vine==1.3.0
whitenoise==4.1.2
